RedApp/requirements.txt
praw
mysql-connector-python
Flask
